refactor(agent): rename codex_tools.rs to coding_tools.rs
The previous filename conflated with two distinct things:

1. The OpenAI Codex product (which the tools mirror by intent — see
   UPCR-2026-020 "Codex Tool Parity" — but the parity is behavioral,
   not a permanent product dependency).
2. The local `codex review` CLI used as a merge gate.

Rename the file and module path to align with the `coding.*`
capability namespace (`coding.image_view.v1`, `coding.dynamic_tool_search.v1`,
`coding.image_generation.v1`, etc.) which is the persistent surface.

Wire format unchanged:
- Tool names (`apply_patch`, `exec_command`, etc.) — unchanged
- Structured metadata key `"codex_tool"` — unchanged (model-visible)
- Error kinds (`coding_tool_missing`, `coding_tool_denied`) — unchanged

Only the Rust file path + module name changes. The UPCR-2026-020
identifier remains the source-of-truth contract document.

20/20 coding_tools tests pass; cargo check clean.
