+++ crates/octos-agent/src/tools/coding_tools.rs
